requirements.txt
pandas
yfinance
finnhub-python
openpyxl
pytz

